# lib/api-zod/src/generated/types/auditEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Machinedog Client Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditAction } from "./auditAction";
import type { AuditEventMetadata } from "./auditEventMetadata";

export interface AuditEvent {
  id: number;
  /** @nullable */
  organizationId?: number | null;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  actorOrganizationId?: number | null;
  /** @nullable */
  actorEmail?: string | null;
  /** @nullable */
  actorUserId?: string | null;
  category: string;
  action: AuditAction;
  /** @nullable */
  targetType?: string | null;
  /** @nullable */
  targetId?: string | null;
  /** @nullable */
  metadata?: AuditEventMetadata;
  createdAt: Date;
}