# lib/api-zod/src/generated/types/projectSecret.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Machinedog Client Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectSecretEnvironment } from "./projectSecretEnvironment";

export interface ProjectSecret {
  id: number;
  organizationId: number;
  projectId: number;
  name: string;
  environment: ProjectSecretEnvironment;
  valuePreview: string;
  version: number;
  /** @nullable */
  createdBy?: string | null;
  /** @nullable */
  updatedBy?: string | null;
  /** @nullable */
  createdByEmail?: string | null;
  /** @nullable */
  updatedByEmail?: string | null;
  /** @nullable */
  lastUsedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}