# lib/api-zod/src/generated/types/updateProjectSecretBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Machinedog Client Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectSecretBodyEnvironment } from "./updateProjectSecretBodyEnvironment";

export interface UpdateProjectSecretBody {
  /** @minLength 1 */
  name?: string;
  environment?: UpdateProjectSecretBodyEnvironment;
  /** @minLength 1 */
  value?: string;
}